backend/requirements.txt
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
bcrypt==4.0.1
click==8.3.3
colorama==0.4.6
Deprecated==1.3.1
dnspython==2.8.0
ecdsa==0.19.2
email-validator==2.3.0
faiss-cpu==1.13.2
fastapi==0.136.1
greenlet==3.5.0
groq==0.9.0
h11==0.16.0
httpx==0.27.0
idna==3.13
limits==5.8.0
Mako==1.3.12
MarkupSafe==3.0.3
packaging==26.2
passlib==1.7.4
pdfplumber==0.11.1
psycopg2-binary==2.9.12
pyasn1==0.6.3
pydantic==2.13.4
pydantic_core==2.46.4
python-dotenv==1.2.2
python-jose==3.5.0
python-multipart==0.0.27
rsa==4.9.1
sentence-transformers==3.0.1
six==1.17.0
slowapi==0.1.9
SQLAlchemy==2.0.49
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.46.0
wrapt==2.1.2
